+++ cli/search/const.go
@@ -1,7 +1,8 @@
 package search
 
 const (
-	ProgramCe = "tarantool"
-	ProgramEe = "tarantool-ee"
-	ProgramTt = "tt"
+	ProgramCe  = "tarantool"
+	ProgramEe  = "tarantool-ee"
+	ProgramTt  = "tt"
+	ProgramDev = "tarantool-dev"
 )