+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - `tt pack` now skips all `.git` files in packed environment, not only in main directory.
+### Added
+
+- `tt install tarantool-dev`: ability to install tarantool from the local build directory.
 
 ## [1.1.2] - 2023-06-16